Move function to adapter module
